+++ src/test/converter2/issues/gh2524.ts
@@ -1,0 +1,6 @@
+export type Alias = {
+    /** {@link Alias.other} {@link other} */
+    readonly default: string;
+    /** {@link Alias.default} {@link default} */
+    readonly other: string;
+};